refactor for separate files for each filter-related component
